Pending changes exported from your codespace

+++ Core/E-Ticaret.Application/ViewModels/Products/VM_Product_Create.cs
@@ -10,6 +10,6 @@ namespace E_Ticaret.Application.ViewModels.Products
         public int CategoryId { get; set; }
         public string Name { get; set; }
         public double Price { get; set; }
-        public short UnitsInStock { get; set; }
+        public int UnitsInStock { get; set; }
     }
 }
