+++ Infrastructure/E-Ticaret.Persistence/Migrations/ETicaretDbContextModelSnapshot.cs
@@ -121,8 +121,8 @@ namespace ETicaret.Persistence.Migrations
                     b.Property<double>("Price")
                         .HasColumnType("double precision");
 
-                    b.Property<short>("UnitsInStock")
-                        .HasColumnType("smallint");
+                    b.Property<int>("UnitsInStock")
+                        .HasColumnType("integer");
 
                     b.Property<DateTime>("UpdatedDate")
                         .HasColumnType("timestamp with time zone");
